Decimal division

+++ src/Expressions.java
@@ -3,11 +3,10 @@ public class Expressions {
         //expressions: operands + operators
         //values, variables, and constants are operands
         //operators : // +, -, *, /, %, ++, --, ==, !=, >, <, >=, <=, &&, ||, !
-        int friends = 10; //int variable
+        double friends = 10; //double variable
         friends = friends/3; //division operator
         System.out.println("I have " + friends + " friends!");
-        //note that the division operator will return the nearest lesser integer value
-        //Example: 3.90 will return 3, not 4
+        //returns decimal value because of double type
 
    
     }
